# doc/api.yml
openapi: 3.0.3
info:
  title: PROG4 final
  version: "1.0.0"
  description: >
    Grade management API (L1–L3). Auth: session cookie (JSESSIONID) via
    POST /auth/login (JSON) or form POST /web/login. HTTP Basic also accepted
    for API clients. Login identifier is the account email (used as username).
servers:
  - url: /
security:
  - sessionAuth: []
...
paths:
...
  /grades:
    post:
      tags: [Exams & grades]
      summary: >
        Record/update a grade (append-only). Each change inserts a new row.
        reason is mandatory when updating an existing grade. ADMIN anywhere;
        TEACHER only on courses assigned for that semester.
      description: Grade value is between 0 and 20.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateGradeRequest" }
      responses:
        "201":
          description: Grade recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GradeResponse" }
        "400":
          description: Missing reason on update, or value out of [0,20]
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Course not assigned to you for this semester
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /exams/{id}/grades:
    get:
      tags: [Exams & grades]
      summary: Current grades of an exam (ADMIN, or TEACHER if assigned)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current grades (latest version per student)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GradeResponse" }
        "403": { description: Access denied }
  /students/{id}/grades:
    get:
      tags: [Exams & grades]
      summary: Current grades of a student (ADMIN, or STUDENT on himself)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current grades
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GradeResponse" }
        "403": { description: Access denied }
  /exams/{id}/grades/{studentId}/history:
    get:
      tags: [Exams & grades]
      summary: >
        Append-only grade history (all versions with value, reason, author,
        enteredAt). ADMIN anywhere; TEACHER if assigned for that semester.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: studentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Grade versions, oldest to newest
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GradeResponse" }
        "403": { description: Access denied }
  /students/me/grades:
    get:
      tags: [Exams & grades]
      summary: Current grades of the authenticated student (STUDENT)
      responses:
        "200":
          description: Current grades
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GradeResponse" }
        "403": { description: Access denied }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    CreateGradeRequest:
      type: object
      required: [studentId, examId, value]
      properties:
        studentId: { type: string, format: uuid }
        examId: { type: string, format: uuid }
        value:
          type: number
          minimum: 0
          maximum: 20
          description: Grade out of 20
        reason:
          type: string
          nullable: true
          description: Mandatory when updating an existing grade
    GradeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        examId: { type: string, format: uuid }
        value: { type: number }
        enteredAt: { type: string, format: date-time }
        reason: { type: string, nullable: true }
        authorId: { type: string, format: uuid }
...
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, description: HTTP status code }
        error: { type: string, description: HTTP reason phrase }
        message: { type: string, description: Human-readable error message }